refactored common taxo logic out into mixin.
